[ALPS05859614] media: cam: fix UASAN issue-II
[Detail]
refine 64bits shift issue

MTK-Commit-Id: bfc1db64ee186dabb5bd0ed18283e753c92deb8c

CR-Id: ALPS05859614
Feature: [Android Default] Camera Application Basic Functions

+++ drivers/media/platform/mtk-isp/camsys/isp7_1/cam/mtk_cam-raw.c
@@ -1721,7 +1721,7 @@ static int mtk_raw_sd_s_stream(struct v4l2_subdev *sd, int enable)
 			if (!pipe->vdev_nodes[i].enabled)
 				continue;
 			pipe->enabled_dmas |=
-				(unsigned int) (1 << pipe->vdev_nodes[i].desc.dma_port);
+				(1ULL << pipe->vdev_nodes[i].desc.dma_port);
 		}
 	} else {
 		for (i = 0; i < ARRAY_SIZE(raw->devs); i++) {
